+++ phase0/crcCard/Board.md
@@ -1,6 +1,6 @@
 [
   {
-    "name": "Board(entity)",
+    "name": "board.Board(entity)",
     "superclasses": "",
     "subclasses": "",
     "type": "Abstract",
