+++ phase0/crcCard/King.md
@@ -13,7 +13,7 @@
       "ValidMove()"
     ],
     "collaborators": [
-      "Board"
+      "board.Board"
     ]
   }
 ]
