+++ phase0/crcCard/Pawn.md
@@ -13,7 +13,7 @@
       "ValidMove()"
     ],
     "collaborators": [
-      "Board"
+      "board.Board"
     ]
   }
 ]
