+++ phase0/crcCard/Piece.md
@@ -15,7 +15,7 @@
       "boolean ValidMove()"
     ],
     "collaborators": [
-      "Board"
+      "board.Board"
     ]
   }
 ]
