+++ phase0/crcCard/Rook.md
@@ -13,7 +13,7 @@
       "ValidMove()"
     ],
     "collaborators": [
-      "Board"
+      "board.Board"
     ]
   }
 ]
