--- phase0/src/main/AI.java
@@ -1,8 +1,0 @@
-public class AI extends Player{
-
-    public AI(String ID) {
-        super(ID);
-    }
-
-    // Designing a game (especially board game) AI could be difficult, maybe leave it until phase 2
-}
