--- phase0/src/main/Human.java
@@ -1,6 +1,0 @@
-public class Human extends Player{
-
-    public Human(String ID) {
-        super(ID);
-    }
-}
