+++ phase0/src/main/board/BoardInterface.java
@@ -1,9 +1,11 @@
+package board;
 
 public interface BoardInterface {
-    String[][] getBoard(); // not for superchess
+
+    String[][] getBoard();
     void addPiece(String pieceName, int X, int Y);
     String removePiece(int X, int Y);
     boolean isPositionVacant(int X, int Y);
     String getPiece(int X, int Y);
-    void reset(); // not for superchess
+    void reset();
 }
