+++ phase0/src/main/player/AI.java
@@ -1,0 +1,12 @@
+package player;
+
+import player.Player;
+
+public class AI extends Player {
+
+    public AI(String ID) {
+        super(ID);
+    }
+
+    // Designing a game (especially board game) player.AI could be difficult, maybe leave it until phase 2
+}
