+++ phase0/src/main/player/Human.java
@@ -1,0 +1,10 @@
+package player;
+
+import player.Player;
+
+public class Human extends Player {
+
+    public Human(String ID) {
+        super(ID);
+    }
+}
