+++ phase0/src/main/player/Player.java
@@ -1,3 +1,5 @@
+package player;
+
 import piece.Color;
 
 public class Player {
