+++ phase0/src/test/TestAI.java
@@ -1,5 +1,5 @@
 import org.junit.Before;
-import org.junit.Test;
+import player.AI;
 
 import static org.junit.Assert.assertEquals;
 
@@ -8,7 +8,7 @@ public class TestAI {
 
     @Before
     public void setup(){
-        ai = new AI("Computer Player");
+        ai = new AI("Computer player.Player");
     }
 
 }
