+++ phase0/src/test/TestBoardManager.java
@@ -1,3 +1,4 @@
+import board.Board;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,7 +15,7 @@ public class TestBoardManager {
     @Before
     public void setup(){
         bm = new BoardManager();
-        b = new Board();
+        b = new Board(8, 8);
     }
 
     @Test(timeout = 50)
