+++ phase0/src/test/TestHuman.java
@@ -1,5 +1,5 @@
 import org.junit.Before;
-import org.junit.Test;
+import player.Human;
 
 import static org.junit.Assert.assertEquals;
 
@@ -8,7 +8,7 @@ public class TestHuman {
 
     @Before
     public void setup(){
-        h = new Human("Human Player");
+        h = new Human("player.Human player.Player");
     }
 
 }
