+++ phase0/src/test/TestPlayer.java
@@ -1,5 +1,5 @@
 import org.junit.Before;
-import org.junit.Test;
+import player.Player;
 
 import static org.junit.Assert.assertEquals;
